refactor(tools): remove pipe_config_validators and tidy tool imports
